First pass at numericProp parsing
